fix(checker): resolve ambient module star exports
